requirements.txt
matplotlib
numpy
scikit_learn
tensorflow_gpu==1.12.0